*-aes chunkers: factor out the shared 80%, no performance change
rabin-aes, goldilocks-aes and toeplitz-aes differed only in their rolling
universal hash, but each carried its own copy of the AES layer, the
two-lane scan structure, the chunker class and the test suite - about 79%
of the goldilocks/toeplitz code was line-identical to rabin-aes.

Three shared pieces replace the copies:

- phte_core.h: AES-128 key expansion, endian-explicit LE helpers, hardware
  detection, single-block AES helpers and the context base (round keys, EVP
  context, path selection).
- phte_scan.h: the scan itself as a compile-time template. Each kernel
  defines its rolling functions and PH_PREFIX/PH_CTX/PH_ROLL/PH_ROLL2/
  PH_DIGEST64, then includes it, so the rolling code is still expanded
  inline into fully specialized scan loops - no function pointers, no
  cross-kernel calls, identical code generation.
- phte_chunker.pyx/.pxd: ChunkerPHTE, the shared chunker driver (buffering,
  min/max clamping, normalized chunking, scan loop, sparse handling). The
  subclasses implement _scan()/_digest64() as Cython cdef methods, i.e. one
  vtable call per scan span (up to a whole chunk) and per chunk start, never
  per byte.

Likewise for the tests: the behavior every one of these chunkers must have
(kernel equivalence, size distribution, shift resilience, param parsing,
fuzzing) now runs parametrized over all three in phte_chunkers_test.py; the
per-chunker modules keep only their frozen golden chunk points and the
tests of their own hash arithmetic and key derivation.

Net -2335 lines. Each kernel .c is now 84-125 lines (was 523-564) and each
chunker .pyx no longer contains the ~250-line class body.

Verified unchanged behavior and performance:
- all three frozen golden chunk-point tests unchanged, 508 tests pass,
  mypy/black/ruff clean, e2e create+extract with all three chunkers OK;
- C-level A/B of the production scan (same harness, HEAD kernels vs these,
  64 MiB, best of 8): rabin 683.6 -> 682.5, goldilocks 331.8 -> 331.8,
  toeplitz 687.4 -> 694.7 MB/s hw; EVP paths 376.6 -> 376.6, 251 -> 250,
  394.2 -> 394.0 - all within +-1%;
- disassembly of ra_scan keeps the identical AES and table-lookup structure
  (853 vs 850 instructions; the delta is prologue register shuffling caused
  by the changed struct field offsets).

Also excludes *.pxd from the wheel (build-time only, borg's first .pxd).
