+++ .gitignore
@@ -14,6 +14,7 @@ src/borg/chunkers/fastcdc.c
 src/borg/chunkers/rabin_aes.c
 src/borg/chunkers/goldilocks_aes.c
 src/borg/chunkers/toeplitz_aes.c
+src/borg/chunkers/phte_chunker.c
 src/borg/chunkers/reader.c
 src/borg/checksums.c
 src/borg/platform/darwin.c
