+++ setup.py
@@ -61,6 +61,7 @@ rabin_aes_source = "src/borg/chunkers/rabin_aes.pyx"
 rabin_aes_impl_source = "src/borg/chunkers/rabin_aes_impl.c"
 goldilocks_aes_source = "src/borg/chunkers/goldilocks_aes.pyx"
 goldilocks_aes_impl_source = "src/borg/chunkers/goldilocks_aes_impl.c"
+phte_chunker_source = "src/borg/chunkers/phte_chunker.pyx"
 toeplitz_aes_source = "src/borg/chunkers/toeplitz_aes.pyx"
 toeplitz_aes_impl_source = "src/borg/chunkers/toeplitz_aes_impl.c"
 reader_source = "src/borg/chunkers/reader.pyx"
@@ -81,6 +82,7 @@ cython_sources = [
     buzhash_source,
     buzhash64_source,
     fastcdc_source,
+    phte_chunker_source,
     rabin_aes_source,
     goldilocks_aes_source,
     toeplitz_aes_source,
@@ -222,6 +224,7 @@ if not on_rtd:
         Extension("borg.chunkers.buzhash", [buzhash_source], extra_compile_args=cflags),
         Extension("borg.chunkers.buzhash64", [buzhash64_source], extra_compile_args=cflags),
         Extension("borg.chunkers.fastcdc", [fastcdc_source], extra_compile_args=cflags),
+        Extension("borg.chunkers.phte_chunker", [phte_chunker_source], extra_compile_args=cflags),
         Extension("borg.chunkers.rabin_aes", **rabin_aes_ext_kwargs),
         Extension("borg.chunkers.goldilocks_aes", **goldilocks_aes_ext_kwargs),
         Extension("borg.chunkers.toeplitz_aes", **toeplitz_aes_ext_kwargs),
